renaming refactoring of the components
